Creo Direccion e ImagenProducto en model / Ajusto y verifico atributos y relaciones en todos los modelos / Creo DireccionController / Creo DireccionService e ImagenProductoService / CreoDireccionRepository, ImagenProductoRepository, ItemCarritoRepository e ItemPedidoRepository / Ajusto el constructor de ProductoDTO

+++ ecommerce/src/main/java/com/tpo_G2/ecommerce/dto/ProductoDTO.java
@@ -9,7 +9,7 @@ public class ProductoDTO {
     private Long id;
     private String nombre;
     private String descripcion;
-    private Double precio;
+    private Float precio;
     private Integer stock;
     private CategoriaDTO categoria;
 }
